+++ table/block_based/data_block_hash_index_test.cc
@@ -580,9 +580,9 @@ void TestBoundary(InternalKey& ik1, std::string& v1, InternalKey& ik2,
   const bool kSkipFilters = true;
   const bool kImmortal = true;
   ASSERT_OK(ioptions.table_factory->NewTableReader(
-      TableReaderOptions(ioptions, moptions.prefix_extractor.get(), soptions,
-                         internal_comparator, !kSkipFilters, !kImmortal,
-                         level_),
+      TableReaderOptions(ioptions, moptions, moptions.prefix_extractor.get(),
+                         soptions, internal_comparator, !kSkipFilters,
+                         !kImmortal, level_),
       std::move(file_reader), sink->contents().size(), &table_reader));
   // Search using Get()
   ReadOptions ro;